Add support for standalone exe, GUI
Create setup.py using cx_freeze to allow for the flush_matrix script to be
run as a standalone exe file (no need for python installation on the
computer running the file). Add a mockup ui file (built using qt)
to outline the look-and-feel of the flush_matrix GUI for when the program
is refactored down the road.