Modify elemental flush algorithm
Update the way flush cycles are calculated based on elemental values
according to the new algorithm sent by Dwayne Z: we now base flushes on
the total fill size rather than a comparison of the absolute difference
between current and target elemental concentrations.
Add improved logging feedback to facilitate testing of the new function.
Change the name and default values of Dye and Demulse Constants to Dye and
Demulse Cycle Count in order to be more intuitive.